Storage: Added CFG_TEE_FS_PARENT_PATH config option
tee-supplicant needs to create folder which will contain secure files.
In case folder can not be created, tee-supplicat exits.

Sometimes developers have to change this location to another.
For example, Android OS can not use /data folder, because tee-supplicant
is needed for decrypt userdata.img (which is mounted in /data folder).

* Add CFG_TEE_FS_PARENT_PATH config option
* Cleanup secure storage init logic

Change-Id: I2a7d5ff281ce71295e2e0013eeaa1f44b8a03e2d
Signed-off-by: Ruslan Piasetskyi <ruslan.piasetskyi@globallogic.com>